+++ src/main/java/com/telepathicgrunt/commandstructures/SpawnPiecesCommand.java
@@ -1,6 +1,7 @@
 package com.telepathicgrunt.commandstructures;
 
 import com.mojang.brigadier.CommandDispatcher;
+import com.mojang.brigadier.arguments.BoolArgumentType;
 import com.mojang.brigadier.arguments.IntegerArgumentType;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.tree.LiteralCommandNode;
@@ -16,20 +17,22 @@ import net.minecraft.commands.arguments.coordinates.WorldCoordinates;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.level.ServerChunkCache;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.StructureBlock;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.StructureBlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.StructureMode;
+import net.minecraft.world.level.chunk.ChunkAccess;
 import net.minecraft.world.level.chunk.LevelChunk;
+import net.minecraft.world.level.levelgen.structure.templatesystem.StructureManager;
+import net.minecraft.world.level.levelgen.structure.templatesystem.StructureTemplate;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -38,44 +41,50 @@ public class SpawnPiecesCommand {
         String commandString = "spawnpieces";
         String rlArg = "resourcelocationpath";
         String locationArg = "location";
-        String fillerblockArg = "fillerblock";
+        String savepieceArg = "savepieces";
         String floorblockArg = "floorblock";
+        String fillerblockArg = "fillerblock";
         String rowlengthArg = "rowlength";
 
         LiteralCommandNode<CommandSourceStack> source = dispatcher.register(Commands.literal(commandString)
                 .requires((permission) -> permission.hasPermission(2))
                 .then(Commands.argument(rlArg, ResourceLocationArgument.id())
-                        .suggests((ctx, sb) -> SharedSuggestionProvider.suggestResource(templatePathsSuggestions(ctx), sb))
-                        .executes(cs -> {
-                            WorldCoordinates worldCoordinates = new WorldCoordinates(
-                                    new WorldCoordinate(false, cs.getSource().getPosition().x()),
-                                    new WorldCoordinate(false, cs.getSource().getPosition().y()),
-                                    new WorldCoordinate(false, cs.getSource().getPosition().z())
-                            );
-                            spawnPieces(cs.getArgument(rlArg, ResourceLocation.class), worldCoordinates, Blocks.STRUCTURE_VOID.defaultBlockState(), Blocks.BARRIER.defaultBlockState(), 13, cs);
-                            return 1;
-                        })
-                        .then(Commands.argument(locationArg, Vec3Argument.vec3())
-                                .executes(cs -> {
-                                    spawnPieces(cs.getArgument(rlArg, ResourceLocation.class), Vec3Argument.getCoordinates(cs, locationArg), Blocks.STRUCTURE_VOID.defaultBlockState(), Blocks.BARRIER.defaultBlockState(), 13, cs);
-                                    return 1;
-                                })
-                                .then(Commands.argument(fillerblockArg, BlockStateArgument.block())
-                                        .executes(cs -> {
-                                            spawnPieces(cs.getArgument(rlArg, ResourceLocation.class), Vec3Argument.getCoordinates(cs, locationArg), BlockStateArgument.getBlock(cs, fillerblockArg).getState(), Blocks.BARRIER.defaultBlockState(), 13, cs);
-                                            return 1;
-                                        })
-                                        .then(Commands.argument(floorblockArg, BlockStateArgument.block())
-                                                .executes(cs -> {
-                                                    spawnPieces(cs.getArgument(rlArg, ResourceLocation.class), Vec3Argument.getCoordinates(cs, locationArg), BlockStateArgument.getBlock(cs, fillerblockArg).getState(), BlockStateArgument.getBlock(cs, floorblockArg).getState(), 13, cs);
-                                                    return 1;
-                                                })
-                                                .then(Commands.argument(rowlengthArg, IntegerArgumentType.integer())
-                                                        .executes(cs -> {
-                                                            spawnPieces(cs.getArgument(rlArg, ResourceLocation.class), Vec3Argument.getCoordinates(cs, locationArg), BlockStateArgument.getBlock(cs, fillerblockArg).getState(), BlockStateArgument.getBlock(cs, floorblockArg).getState(), cs.getArgument(rowlengthArg, Integer.class), cs);
-                                                            return 1;
-                                                        })
-                                                ))))));
+                .suggests((ctx, sb) -> SharedSuggestionProvider.suggestResource(templatePathsSuggestions(ctx), sb))
+                .executes(cs -> {
+                    WorldCoordinates worldCoordinates = new WorldCoordinates(
+                            new WorldCoordinate(false, cs.getSource().getPosition().x()),
+                            new WorldCoordinate(false, cs.getSource().getPosition().y()),
+                            new WorldCoordinate(false, cs.getSource().getPosition().z())
+                    );
+                    spawnPieces(cs.getArgument(rlArg, ResourceLocation.class), worldCoordinates, false, Blocks.BARRIER.defaultBlockState(), Blocks.AIR.defaultBlockState(), 13, cs);
+                    return 1;
+                })
+                .then(Commands.argument(locationArg, Vec3Argument.vec3())
+                .executes(cs -> {
+                    spawnPieces(cs.getArgument(rlArg, ResourceLocation.class), Vec3Argument.getCoordinates(cs, locationArg), false, Blocks.BARRIER.defaultBlockState(), Blocks.AIR.defaultBlockState(), 13, cs);
+                    return 1;
+                })
+                .then(Commands.argument(savepieceArg, BoolArgumentType.bool())
+                .executes(cs -> {
+                    spawnPieces(cs.getArgument(rlArg, ResourceLocation.class), Vec3Argument.getCoordinates(cs, locationArg), cs.getArgument(savepieceArg, boolean.class), Blocks.BARRIER.defaultBlockState(), Blocks.AIR.defaultBlockState(), 13, cs);
+                    return 1;
+                })
+                .then(Commands.argument(floorblockArg, BlockStateArgument.block())
+                .executes(cs -> {
+                    spawnPieces(cs.getArgument(rlArg, ResourceLocation.class), Vec3Argument.getCoordinates(cs, locationArg), cs.getArgument(savepieceArg, boolean.class), BlockStateArgument.getBlock(cs, floorblockArg).getState(), Blocks.AIR.defaultBlockState(), 13, cs);
+                    return 1;
+                })
+                .then(Commands.argument(fillerblockArg, BlockStateArgument.block())
+                .executes(cs -> {
+                    spawnPieces(cs.getArgument(rlArg, ResourceLocation.class), Vec3Argument.getCoordinates(cs, locationArg), cs.getArgument(savepieceArg, boolean.class), BlockStateArgument.getBlock(cs, floorblockArg).getState(), BlockStateArgument.getBlock(cs, fillerblockArg).getState(), 13, cs);
+                    return 1;
+                })
+                .then(Commands.argument(rowlengthArg, IntegerArgumentType.integer())
+                .executes(cs -> {
+                    spawnPieces(cs.getArgument(rlArg, ResourceLocation.class), Vec3Argument.getCoordinates(cs, locationArg), cs.getArgument(savepieceArg, boolean.class), BlockStateArgument.getBlock(cs, floorblockArg).getState(), BlockStateArgument.getBlock(cs, fillerblockArg).getState(), cs.getArgument(rowlengthArg, Integer.class), cs);
+                    return 1;
+                })
+        )))))));
 
         dispatcher.register(Commands.literal(commandString).redirect(source));
     }
@@ -106,30 +115,29 @@ public class SpawnPiecesCommand {
         return rlSet;
     }
 
-    public static void spawnPieces(ResourceLocation path, Coordinates coordinates, BlockState fillBlockState, BlockState floorBlockState, int rowlength, CommandContext<CommandSourceStack> cs) {
+    public static void spawnPieces(ResourceLocation path, Coordinates coordinates, boolean savePieces, BlockState floorBlockState, BlockState fillBlockState, int rowlength, CommandContext<CommandSourceStack> cs) {
         ServerLevel level = cs.getSource().getLevel();
         Player player = cs.getSource().getEntity() instanceof Player player1 ? player1 : null;
         BlockPos pos = coordinates.getBlockPos(cs.getSource());
 
-        List<ResourceLocation> identifiers = getResourceLocations(player, level, path.getNamespace(), path.getPath());
+        List<ResourceLocation> nbtRLs = getResourceLocations(player, level, path.getNamespace(), path.getPath());
 
         // Size of area we will need
         int columnCount = rowlength;
-        int rowCount = (int) Math.max(Math.ceil((float)identifiers.size() / columnCount), 1);
+        int rowCount = (int) Math.max(Math.ceil((float)nbtRLs.size() / columnCount), 1);
         if(rowCount == 1) {
-            columnCount = identifiers.size();
+            columnCount = nbtRLs.size();
         }
 
         int spacing = 48;
         BlockPos bounds = new BlockPos((spacing * rowCount) + 16, spacing, spacing * columnCount);
 
         // Fill/clear area with structure void
-        BlockPos.MutableBlockPos mutableChunk = clearAreaNew(level, pos, player, bounds, fillBlockState, floorBlockState);
-
-        generateStructurePieces(level, pos, player, identifiers, columnCount, spacing, mutableChunk);
+        clearAreaNew(level, pos, player, bounds, fillBlockState, floorBlockState);
+        generateStructurePieces(level, pos, player, nbtRLs, columnCount, spacing, savePieces);
     }
 
-    private static BlockPos.MutableBlockPos clearAreaNew(Level world, BlockPos pos, Player player, BlockPos bounds, BlockState fillBlock, BlockState floorBlock) {
+    private static void clearAreaNew(ServerLevel world, BlockPos pos, Player player, BlockPos bounds, BlockState fillBlock, BlockState floorBlock) {
         BlockPos.MutableBlockPos mutableChunk = new BlockPos.MutableBlockPos().set(pos.getX() >> 4, pos.getY(), pos.getZ() >> 4);
         mutableChunk.move(1,0,0);
         int endChunkX = (pos.getX() + bounds.getX()) >> 4;
@@ -150,14 +158,14 @@ public class SpawnPiecesCommand {
                         mutable.setY(pos.getY());
                         BlockState oldState = chunk.setBlockState(mutable, floorBlock, false);
                         if(oldState != null) {
-                            ((ServerChunkCache)world.getChunkSource()).blockChanged(mutable);
+                            world.getChunkSource().blockChanged(mutable);
                             world.getChunkSource().getLightEngine().checkBlock(mutable);
                         }
                         for(int y = pos.getY() + 1; y < pos.getY() + 64; y++) {
                             mutable.setY(y);
                             oldState = chunk.setBlockState(mutable, fillBlock, false);
                             if(oldState != null) {
-                                ((ServerChunkCache)world.getChunkSource()).blockChanged(mutable);
+                                world.getChunkSource().blockChanged(mutable);
                                 world.getChunkSource().getLightEngine().checkBlock(mutable);
                             }
                         }
@@ -170,7 +178,6 @@ public class SpawnPiecesCommand {
             }
             mutableChunk.set(mutableChunk.getX(), mutableChunk.getY(), pos.getZ() >> 4); // Set back to start of row
         }
-        return mutableChunk;
     }
 
 
@@ -189,36 +196,66 @@ public class SpawnPiecesCommand {
     }
 
 
-    private static void generateStructurePieces(Level world, BlockPos pos, Player player, List<ResourceLocation> identifiers, int columnCount, int spacing, BlockPos.MutableBlockPos mutableChunk) {
-        mutableChunk.set(((pos.getX() >> 4) + 1) << 4, pos.getY(), (pos.getZ() >> 4) << 4);
+    private static void generateStructurePieces(ServerLevel world, BlockPos pos, Player player, List<ResourceLocation> nbtRLs, int columnCount, int spacing, boolean savePieces) {
+        BlockPos.MutableBlockPos mutable = new BlockPos.MutableBlockPos().set(((pos.getX() >> 4) + 1) << 4, pos.getY(), (pos.getZ() >> 4) << 4);
 
-        for(int pieceIndex = 1; pieceIndex <= identifiers.size(); pieceIndex++) {
+        for(int pieceIndex = 1; pieceIndex <= nbtRLs.size(); pieceIndex++) {
             if(player != null) {
-                player.displayClientMessage(new TranslatableComponent(" Working making structure: " + identifiers.get(pieceIndex - 1)), true);
+                player.displayClientMessage(new TranslatableComponent(" Working making structure: " + nbtRLs.get(pieceIndex - 1)), true);
             }
 
-            world.setBlock(mutableChunk, Blocks.STRUCTURE_BLOCK.defaultBlockState().setValue(StructureBlock.MODE, StructureMode.LOAD), 3);
-            BlockEntity be = world.getBlockEntity(mutableChunk);
+            world.setBlock(mutable, Blocks.STRUCTURE_BLOCK.defaultBlockState().setValue(StructureBlock.MODE, StructureMode.LOAD), 3);
+            BlockEntity be = world.getBlockEntity(mutable);
             if(be instanceof StructureBlockEntity structureBlockTileEntity) {
-                structureBlockTileEntity.setStructureName(identifiers.get(pieceIndex-1)); // set identifier
+                structureBlockTileEntity.setStructureName(nbtRLs.get(pieceIndex-1)); // set identifier
 
                 structureBlockTileEntity.setMode(StructureMode.LOAD);
                 structureBlockTileEntity.setIgnoreEntities(false);
-                structureBlockTileEntity.loadStructure((ServerLevel) world,false); // load structure
+
+                fillStructureVoidSpace(world, nbtRLs.get(pieceIndex-1), mutable);
+                structureBlockTileEntity.loadStructure(world,false); // load structure
 
                 structureBlockTileEntity.setMode(StructureMode.SAVE);
-                //structureBlockTileEntity.saveStructure(true); //save structure
+                if(savePieces) {
+                    structureBlockTileEntity.saveStructure(true);
+                }
                 //structureBlockTileEntity.setShowAir(true);
                 structureBlockTileEntity.setIgnoreEntities(false);
             }
 
-            mutableChunk.move(0,0, spacing);
+            mutable.move(0,0, spacing);
 
 
             // Move back to start of row
             if(pieceIndex % columnCount == 0) {
-                mutableChunk.move(spacing,0, (-spacing * columnCount));
+                mutable.move(spacing,0, (-spacing * columnCount));
             }
         }
+    }
+
+    // Needed so that structure void is preserved in structure pieces.
+    private static void fillStructureVoidSpace(ServerLevel world, ResourceLocation resourceLocation, BlockPos startSpot) {
+        StructureManager structuremanager = world.getStructureManager();
+        Optional<StructureTemplate> optional = structuremanager.get(resourceLocation);
+        optional.ifPresent(template -> {
+            BlockPos.MutableBlockPos mutable = startSpot.mutable();
+            ChunkAccess chunk = world.getChunk(mutable);
+            for(int x = 0; x < template.getSize().getX(); x++) {
+                for (int z = 0; z < template.getSize().getZ(); z++) {
+                    for(int y = 0; y < template.getSize().getY(); y++) {
+                        mutable.set(startSpot).move(x, y + 1, z);
+                        if(chunk.getPos().x != mutable.getX() >> 4 || chunk.getPos().z != mutable.getZ() >> 4) {
+                            chunk = world.getChunk(mutable);
+                        }
+
+                        BlockState oldState = chunk.setBlockState(mutable, Blocks.STRUCTURE_VOID.defaultBlockState(), false);
+                        if(oldState != null) {
+                            world.getChunkSource().blockChanged(mutable);
+                            world.getChunkSource().getLightEngine().checkBlock(mutable);
+                        }
+                    }
+                }
+            }
+        });
     }
 }